DIP creation based off DIP directory, not unitVars
refs #6868

Approve normalization had two MicroServiceChainLinks nodes, which
caused problems since the new processingMCP matches based on UUID not
node name.  However, the only reason to have two approve nodes was to
set whether or not a DIP was created.

Update the DIP creation to check to be based on whether or not the DIP
directory exists, instead of unitVars controlling the workflow.  Remove
all related workflow unitVars.  Merge approve normalization nodes.