AIP Reingest: DC metadata marked when updated
refs #6488

DC metadata should be written to the METS file only if it has been
changed during reingest. To track this, add a status field to the
DublinCore table which can be ORIGINAL, REINGEST or UPDATED.  ORIGINAL
is the default for AIPs being created normally, REINGEST is the status
when it is initially parsed to the DB, UPDATED is when the data is
changed after reingest.

Fix DublinCore.metadataappliestotype to be a foreign key.