Update PREMIS Rights models & forms for mandatory/optional fields
refs #6488

Update blank & null attributes on PREMIS Rights related tables to reflect
which fields were mandatory or optional in the PREMIS spec.  If a field is
optional, but in its own table (eg RightsStatementCopyrightNote), then it
is marked as blank=null=False, since that row will only be created if that
field has information.  Update SQL for those tables to reflect the Django
model.