Fix missing structMap, TYPE attrs
refs #7068, #7071

Remove the FileLevelOfDescription and replace it with attributes on
SIPArrange. Instead of attaching the level of description to the
FileLevelOfDescription, populate the new level_of_description field in
SIPArrange. This is easier to find, set and update when the SIP is created,
and less error prone when creating a SIP from not top level directories. Do
not delete SIPArrange directories, as they are now needed for level of
description information.

In createMETS2, search based on the path relative to the objects directory.
Query SIPArrange instead of FileLevelOfDescription.