make unit_type an optional parameter when deleting files
refs #7357

the connect_and_remove_transfer_files method requires two parameters,
the uuid of the transfer or sip that you want to remove from the
ElasticSearch index, and the unit_type, which can be either
'sip' or 'transfer'.

This change makes the unit_type parameter optional (default is None)
that way it is not necessary to know if the package you want to remove
is a sip or a transfer.  The uuid uniquely identifies it, so
the unit_type does not help guarantee a match in any case.