Pair matcher: correctly page results
Previous to this, the Archivists' Toolkit code would fetch all results
from the database. This is relatively wasteful, but is *much* slower in
ArchivesSpace due to the large numbers of REST calls.

This instead fetches as few records as possible in order to improve
speed. In order to allow the Django Paginator to still be used, a new
list-like type has been introduced which provides a list-like interface
to paged data sources like REST clients.